+++ R/print.R
@@ -228,6 +228,36 @@ print.tna_cliques <- function(x, n = 6, first = 1,
   invisible(x)
 }
 
+#' Print a TNA data object
+#'
+#' @export
+#' @param x A `tna_data` object.
+#' @param format A `character` string that defines the format of the printed
+#' tibble. Accepts either `"wide"` (default) for wide format data or `"long"`
+#' for long format.
+#' @param ... Arguments passed to the tibble `print` method.
+#' @return `x` (invisibly).
+#' @examples
+#' data_single_session <- tibble::tibble(
+#'   action = c(
+#'     "view", "click", "add_cart", "view", "checkout", "view", "click", "share"
+#'    )
+#' )
+#' results_single <- prepare_data(data_single_session, action = "action")
+#' print(results_single, format = "wide")
+#' print(results_single, format = "long")
+#'
+print.tna_data <- function(x, format = "wide", ...) {
+  check_missing(x)
+  check_class(x, "tna_data")
+  format <- check_match(format, c("wide", "long"))
+  if (format == "wide") {
+    print(x$wide_format, ...)
+  } else {
+    print(x$long_format, ...)
+  }
+}
+
 #' Print Centrality Stability Results
 #'
 #' @export